requirements.txt
# Core AI and ML
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
pinecone>=3.0.0

# Web Framework
Flask>=3.0.0
flask-socketio>=5.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Document Processing
pypdf>=4.0.0
openpyxl>=3.1.0

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1

# Data Processing
pandas>=2.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Async Processing
gevent>=23.0.0

# Dashboard and Visualization
streamlit>=1.28.0

# Utilities
colorama>=0.4.6

# Additional Dependencies for Sentiment Analysis
nltk>=3.9.0
joblib>=1.3.0

# WebSocket Support
python-socketio>=5.14.0
python-engineio>=4.12.0

# Binance API
python-binance>=1.0.19

# HTTP client (Z-API WhatsApp)
requests>=2.28.0